Added isFull() method to Column. Added a UtilityFunction interfaced and modified BasicUtilityFunction so that it implements it. Wrote the buildTree method for TreeWorker (not yet tested) Added a test file for BasicUtilityFunction. Added a test file for TreeWorker.

+++ ConnectN/src/edu/wpi/cs/connectn/tree/Node.java
@@ -16,7 +16,7 @@ public class Node {
 	protected final Move move;
 	
 	/** The utility of being at this node (or game state) */
-	protected int utility;
+	protected float utility;
 	
 	/** The parent node of this node */
 	protected final Node parent;
@@ -24,6 +24,13 @@ public class Node {
 	/** The depth of this node in the tree */
 	protected final int depth;
 	
+	/**
+	 * Constructs a new Node with the given Move, parent node, utility, and depth
+	 * @param move the move this node will represent
+	 * @param parent the parent of this node
+	 * @param utility the utility of being at the state represented by this node
+	 * @param depth the depth in the tree of this node
+	 */
 	public Node(Move move, Node parent, int utility, int depth) {
 		this.move = move;
 		this.parent = parent;
@@ -80,14 +87,14 @@ public class Node {
 	/**
 	 * @return the utility
 	 */
-	public int getUtility() {
+	public float getUtility() {
 		return utility;
 	}
 
 	/**
 	 * @param utility the utility to set
 	 */
-	public void setUtility(int utility) {
+	public void setUtility(float utility) {
 		this.utility = utility;
 	}
